requirements.txt
git+https://github.com/dpallot/simple-websocket-server.git
pyserial
